+++ CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.12.2] - 2021-09-09
+## [1.13.2] - 2021-09-24
 
 ### Updated
 
 - Documentation updated to include event versioning guide.
+
+## [1.13.1] - 2021-09-10
+
+### Fixed
+
+- Fixed unit tests
+
+## [1.13.0] - 2021-08-27
+
+### Added
+
+- Messages to handle MODAL open and close event
+- Add `isInsideShellModal()` method to know if app is running inside Shell Modal
 
 ## [1.12.1] - 2021-08-11